Integer type too small for user.id

+++ twitoff_src/models.py
@@ -5,7 +5,7 @@ DB = SQLAlchemy()
 
 class User(DB.Model):
     """Twitter users that we pull and analyze Tweets for."""
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     name = DB.Column(DB.String(20), nullable=False)
     newest_tweet_id = DB.Column(DB.BigInteger)
